Break colors into it's own file
